pyproject.toml
[project]
name = "audio-transcriptor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "audio-recorder-streamlit>=0.0.10",
    "groq>=0.26.0",
    "langchain>=0.3.25",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.7",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
